% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgData.R
\docType{data}
\name{imgData}
\alias{imgData}
\title{Imagery data over a forest area}
\format{An object of class RasterLayer.}
\usage{
data(imgData)
}
\description{
A dataset containing a band (at 812 nm) of an hyperspectral image acquired over a forest area.
}
\details{
\itemize{
  \item imgData An object of class RasterLayer.
}
}
\keyword{datasets}
