% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasData.R
\docType{data}
\name{lasData}
\alias{lasData}
\title{LiDAR data point cloud acquired over a forest area}
\format{A data frame with 16907 rows and 3 variables}
\usage{
data(lasData)
}
\description{
A dataset containing the X Y Z coordinates of LiDAR points acquired over a forest area. The EPSG code of the coordinates is 32632.
}
\details{
\itemize{
  \item X. X coordinate in UTM WGS84 32 N (EPSG code 32632).
  \item Y. Y coordinate in UTM WGS84 32 N (EPSG code 32632).
  \item Z. Z coordinate. The Z coordinate is normalized respect to the ground.
}
}
\keyword{datasets}

